feat: update
Updated handlers and helper functions to work with Redis
Added logging levels in thin areas

+++ utils/utils.py
@@ -26,30 +26,43 @@ class MessageProcessor:
         Retrieves the set of message IDs to delete from the state using the
         provided key, attempts to delete each message,
         and updates the state to clear the set of message IDs.
-
         Logs the start and end of the deletion process, and any errors
         encountered.
-
         :param key: Str
         :return: None
         """
-        logger_utils.debug('Starting to delete messages…')
-        logger_utils.debug(f'Вход\n{self._state.__dict__}')
+        logger_utils.debug(f'Вход')
+        try:
+            msgs: list = dict(await self._state.get_data()).get(key, [])
+        except Exception as e:
+            logger_utils.error(f'Error getting data: {e}')
+            msgs = []
+        logger_utils.debug(f'{msgs=}')
 
-        data: set = dict(await self._state.get_data()).get(key, set())
         if isinstance(self._type_update, Message):
             chat_id = self._type_update.chat.id
         else:
             chat_id = self._type_update.message.chat.id
 
-        for msg_id in data:
-            try:
-                await self._type_update.bot.delete_message(chat_id=chat_id,
-                                                           message_id=msg_id)
-            except TelegramBadRequest as err:
-                logger_utils.warning(f'Failed to remove inline keyboard: {err}')
-        await self._state.update_data({key: set()})
-        logger_utils.debug('Messages deleted')
+        if msgs:
+            logger_utils.debug(f'Starting to delete messages…')
+            for msg_id in set(msgs):
+                try:
+                    await self._type_update.bot.delete_message(chat_id=chat_id,
+                                                               message_id=msg_id)
+                    logger_utils.debug(f'Message with id {msg_id} deleted')
+                except TelegramBadRequest as err:
+                    if "message to delete not found" in str(err):
+                        logger_utils.warning(
+                            f'Failed to delete message with id {msg_id}: {err}')
+                    else:
+                        logger_utils.error(
+                            f'Failed to delete message with id {msg_id}: {err}')
+            # Clearing the storage after successfully deleting all messages
+            await self._state.update_data({key: []})
+        else:
+            logger_utils.debug('No data to delete.')
+        logger_utils.debug('Выход')
 
     async def writes_msg_id_to_storage(
             self, value: Message | CallbackQuery, key='msg_for_del') -> None:
@@ -67,8 +80,8 @@ class MessageProcessor:
         """
         logger_utils.debug('Start writing data to storage…')
 
-        data: set[int] = dict(await self._state.get_data()).get(key, set())
-        data.add(value.message_id)
+        data: list = dict(await self._state.get_data()).get(key, [])
+        data.append(value.message_id)
         await self._state.update_data({key: data})
 
         logger_utils.debug('Message ID to recorded')
@@ -88,20 +101,21 @@ class MessageProcessor:
         """
         logger_utils.debug('Starting delete keyboard…')
 
-        data: set = dict(await self._state.get_data()).get(key, set())
+        msgs: list = dict(await self._state.get_data()).get(key, [])
+
         if isinstance(self._type_update, Message):
             chat_id = self._type_update.chat.id
         else:
             chat_id = self._type_update.message.chat.id
 
-        for msg_id in data:
+        for msg_id in set(msgs):
             try:
                 await self._type_update.bot.edit_message_reply_markup(
                         chat_id=chat_id, message_id=msg_id)
             except TelegramBadRequest as err:
                 logger_utils.error(f'Failed to remove inline keyboard: {err}',
                                    exc_info=True)
-        await self._state.update_data({key: set()})
+        await self._state.update_data({key: []})
 
         logger_utils.debug('Keyboard removed')
